+++ example/lib/main.dart
@@ -51,7 +51,6 @@ class _MyHomePageState extends State<MyHomePage> {
         body: Padding(
           padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 16),
           child: NRCField(
-            language: NrcLanguage.myanmar,
             onCompleted: (value) {
               debugPrint("onCompleted : $value");
             },